[ALPS03103843] USB: fix 32-bit build error
[Detail]
add type casting method by "uintptr_t"

MTK-Commit-Id: 0e28cdb17224a0b830cc8f980210e5e30e629609

CR-Id: ALPS03103843
Feature: USB

+++ drivers/misc/mediatek/usb20/musb_host.c
@@ -541,7 +541,7 @@ static void musb_advance_schedule(struct musb *musb, struct urb *urb, struct mus
 	u64 urb_val;
 
 	/* special case to handle QH memory leak */
-	urb_val = (u64)urb;
+	urb_val = (u64)(uintptr_t)urb;
 	switch (urb_val) {
 	case QH_FREE_RESCUE_INTERRUPT:
 	case QH_FREE_RESCUE_EP_DISABLE:
